# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Curriculum Activity & Attendance App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: month
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: section
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceListResponse"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Attendance marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/bulk:
    post:
      operationId: markBulkAttendance
      tags: [attendance]
      summary: Mark attendance for multiple students at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAttendanceInput"
      responses:
        "201":
          description: Bulk attendance marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAttendanceResult"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
...
  /attendance/low-attendance:
    get:
      operationId: getLowAttendanceStudents
      tags: [attendance]
      summary: Get students with attendance below 75%
      parameters:
        - name: threshold
          in: query
          schema:
            type: integer
            default: 75
      responses:
        "200":
          description: Students with low attendance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentAttendanceInfo"

  # TIMETABLE
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activity recommendations
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create an activity recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # NOTIFICATIONS
...
  /analytics/daily:
    get:
      operationId: getDailyAnalytics
      tags: [analytics]
      summary: Daily attendance chart data
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyAnalytics"
  /analytics/weekly:
    get:
      operationId: getWeeklyAnalytics
      tags: [analytics]
      summary: Weekly attendance chart data
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: weeks
          in: query
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Weekly analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyAnalytics"
  /analytics/monthly:
    get:
      operationId: getMonthlyAnalytics
      tags: [analytics]
      summary: Monthly attendance chart data
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyAnalytics"
...
components:
  schemas:
...
    AttendanceRecord:
      type: object
      required: [id, studentId, subjectId, date, status, method]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        registerNumber:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late]
        method:
          type: string
          enum: [manual, qr, face]
        markedAt:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
    AttendanceInput:
      type: object
      required: [studentId, subjectId, date, status, method]
      properties:
        studentId:
          type: integer
        subjectId:
          type: integer
        date:
          type: string
          format: date
        status:
          type: string
          enum: [present, absent, late]
        method:
          type: string
          enum: [manual, qr, face]
        notes:
          type: string
    AttendanceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [present, absent, late]
        notes:
          type: string
    BulkAttendanceInput:
      type: object
      required: [subjectId, date, method, records]
      properties:
        subjectId:
          type: integer
        date:
          type: string
          format: date
        method:
          type: string
          enum: [manual, qr, face]
        records:
          type: array
          items:
            type: object
            required: [studentId, status]
            properties:
              studentId:
                type: integer
              status:
                type: string
                enum: [present, absent, late]
    BulkAttendanceResult:
      type: object
      required: [created, updated, total]
      properties:
        created:
          type: integer
        updated:
          type: integer
        total:
          type: integer
    AttendanceListResponse:
      type: object
      required: [records, total, page, limit]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    StudentAttendanceInfo:
      type: object
      required: [studentId, name, registerNumber, departmentName, percentage]
      properties:
        studentId:
          type: integer
        name:
          type: string
        registerNumber:
          type: string
        departmentName:
          type: string
        year:
          type: integer
        section:
          type: string
        percentage:
          type: number
...
    Activity:
      type: object
      required: [id, title, category, description]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: string
        resourceUrl:
          type: ["string", "null"]
        difficulty:
          type: ["string", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        isActive:
          type: boolean
    ActivityInput:
      type: object
      required: [title, category, description]
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        resourceUrl:
          type: string
        difficulty:
          type: string
        estimatedMinutes:
          type: integer
        isActive:
          type: boolean
    ActivityUpdate:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        resourceUrl:
          type: string
        difficulty:
          type: string
        estimatedMinutes:
          type: integer
        isActive:
          type: boolean
...
    DailyAnalytics:
      type: object
      required: [date, present, absent, late, total, percentage]
      properties:
        date:
          type: string
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        total:
          type: integer
        percentage:
          type: number
    WeeklyAnalytics:
      type: object
      required: [week, present, absent, total, percentage]
      properties:
        week:
          type: string
        present:
          type: integer
        absent:
          type: integer
        total:
          type: integer
        percentage:
          type: number
    MonthlyAnalytics:
      type: object
      required: [month, present, absent, total, percentage]
      properties:
        month:
          type: string
        present:
          type: integer
        absent:
          type: integer
        total:
          type: integer
        percentage:
          type: number
...