# lib/api-zod/src/generated/types/dailyAnalytics.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Curriculum Activity & Attendance App API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyAnalytics {
  date: string;
  present: number;
  absent: number;
  late: number;
  total: number;
  percentage: number;
}